compile passed but has run time error (vector index out of range...)
