+++ AnalysisTableItem.cpp
@@ -1,0 +1,10 @@
+#pragma once
+#include "stdafx.h"
+#include "AnalysisTableItem.hpp"
+
+//implements
+AnalysisTableItem::AnalysisTableItem(const AnalysisTableItem& a){
+	actionType = a.actionType;
+	stateNumber = a.stateNumber;
+	semanticActionID = a.semanticActionID;
+};
